+++ currency-trends-monitor-server/src/main/java/kae/demo/currencytrendsmonitor/server/CurrencyTrendsMonitorServerApplication.java
@@ -1,4 +1,4 @@
-package kae.demo.currencytrendsmonitorserver;
+package kae.demo.currencytrendsmonitor.server;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
