+++ currency-trends-monitor-server/src/test/java/kae/demo/currencytrendsmonitor/server/CurrencyTrendsMonitorServerApplicationTests.java
@@ -1,4 +1,4 @@
-package kae.demo.currencytrendsmonitorserver;
+package kae.demo.currencytrendsmonitor.server;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
